refactor(color-picker): renamed `ColorInput` to `ColorSelector`
