+++ .changeset/four-spies-run.md
@@ -1,0 +1,5 @@
+---
+"@yamada-ui/color-picker": patch
+---
+
+Renamed `ColorInput` to `ColorPicker` and `ColorPicker` to `ColorSelector`.
