+++ packages/components/color-picker/src/index.ts
@@ -1,7 +1,7 @@
+export { ColorSelector } from "./color-selector"
 export { ColorPicker } from "./color-picker"
-export { ColorInput } from "./color-input"
+export type { ColorSelectorProps } from "./color-selector"
 export type { ColorPickerProps } from "./color-picker"
-export type { ColorInputProps } from "./color-input"
 export { HueSlider } from "./hue-slider"
 export type { HueSliderProps } from "./hue-slider"
 export { AlphaSlider } from "./alpha-slider"
